database now initialised in index
