+++ index.js
@@ -1,5 +1,17 @@
 const CLI = require('./lib/cli.js');
+const mysql = require('mysql2');
+require('dotenv').config();
 
-const cli = new CLI();
+const db = mysql.createConnection(
+    {
+        host: 'localhost',
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_NAME,
+    },
+    console.log("Connected to Database")
+)
 
-cli.run();
+const cli = new CLI(db);
+
+cli.mainMenu();
