+++ lib/cli.js
@@ -1,34 +1,16 @@
 // Imports
 
-const mysql = require('mysql2');
+// const mysql = require('mysql2');
 const inquirer = require('inquirer');
 const { mainMenu, addDepartment, addRole, addEmployee, updateEmployee } = require('./questions');
 const { Departments, Employees } = require('../db/employeeDatabase.js')
-require('dotenv').config();
-
-const db = mysql.createConnection(
-    {
-        host: 'localhost',
-        user: process.env.DB_USER,
-        password: process.env.DB_PASSWORD,
-        database: process.env.DB_NAME,
-    },
-    console.log("Connected to Database")
-)
-
 
 class CLI {
 
-    constructor() {
-
+    constructor(db) {
+        this.db = db
     }
 
-    // Function to connect to the database.
-    run() {
-        db.connect()
-        this.mainMenu()
-    }
-          // TAKE RESPONSE => RUN CLASS
     mainMenu() {
         inquirer
         .prompt(mainMenu)
@@ -38,7 +20,7 @@ class CLI {
                 viewEmployees: Employees
             }
 
-            const data = new option[response.option](db);
+            const data = new option[response.option](this.db);
             return data.retrieve();
         })
         .then((data) => console.table(data))
